tried changing long to BigDecimal in price

+++ src/test/java/wishlist/integration/WishlistRepositoryTest.java
@@ -10,6 +10,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.jdbc.Sql;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -122,7 +123,7 @@ public class WishlistRepositoryTest {
 
         String newNote = "Updated note";
         String newUrl = "http://updated-url.com";
-        long newPrice = 12345;
+        BigDecimal newPrice = 12345;
 
         int rows = wishlistRepository.updateWishlistItem(wishlistId, itemId, newNote, newUrl, newPrice);
 
@@ -169,7 +170,7 @@ public class WishlistRepositoryTest {
 
         String note = "JUnit note";
         String url = "http://example.com";
-        long price = 999;
+        BigDecimal price = 999;
 
         int rows = wishlistRepository.addItemToWishlist(wishlistId, newItemId, note, url, price);
